통합 테스트

See merge request s09-webmobile2-sub2/S09P12C210!124

+++ backend/src/main/java/com/kkini/core/domain/scrap/service/ScrapServiceImpl.java
@@ -37,9 +37,13 @@ public class ScrapServiceImpl implements ScrapService{
     }
 
     @Override
-    public void deleteScrap(Long id) {
+    public void deleteScrap(Long id, Long memberId) {
         Scrap scrap = scrapRepository.findById(id).orElseThrow(() -> new NotFoundException(Scrap.class, id));
-        scrapRepository.delete(scrap);
+        Member member = memberRepository.findById(memberId).orElseThrow(() -> new NotFoundException(Member.class, memberId));
+
+        if (scrap.getMember().equals(member)) {
+            scrapRepository.delete(scrap);
+        }
     }
 
     @Override
